+++ examples/binutils.py
@@ -21,7 +21,8 @@ def main() -> None:
 
     test_cases: list[TestCase] = []
 
-    with open("examples/binutils_gdb_bugs/bugs.json", "r") as f:
+    prefix: str = os.path.join("examples", "bugs", "binutils")
+    with open(os.path.join(prefix, "bugs.json"), "r") as f:
         data: Any = json.load(f)
         for bug in data:
             if bug["skip"]:
@@ -29,7 +30,7 @@ def main() -> None:
 
             test_cases.append(
                 TestCase.make_file(
-                    input_file=bug["file"],
+                    input_file=os.path.join(prefix, bug["file"]),
                     output_file=os.path.join("/tmp", os.path.basename(bug["file"])),
                     algorithms=[DDMin(), ZipMin(), ProbDD()],
                     command=bug["command"],
@@ -41,8 +42,8 @@ def main() -> None:
                 )
             )
 
-    benchmark: Benchmark = Benchmark(test_cases, "/tmp/results.json")
-    validates: list[bool] = benchmark.validate()
+    benchmark: Benchmark = Benchmark(test_cases, os.path.join("/tmp", "results.json"))
+    validates: list[bool] = benchmark.validate(show_process=True)
     if not all(validates):
         print(validates)
         print("Some test cases are invalid. Please check the environment.")