+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimized the `Configuration` class.
 - Updated `ZipMin` class for better performance.
 - Updated `install.sh`.
+- Updated examples to match new API.
 
 ## [0.1.0] - 2025-10-11